requirements.txt
asgiref==3.5.2
attrs==22.1.0
bidict==0.22.0
coverage==6.5.0
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
flake8==5.0.4
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.4
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.4
sqlparse==0.4.3
tomli==2.0.1
